[Silver V] Title: 공포의 면담실, Time: 600 ms, Memory: 78736 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/30088. 공포의 면담실/공포의 면담실.java
@@ -8,16 +8,16 @@ public class Main {
 	public static void main(String[] args) throws IOException {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		int N = Integer.parseInt(br.readLine()); // 부서의 수
-		int[] arr = new int[N];
+		long[] arr = new long[N];
 		for (int i = 0; i < N; i++) {
 			st = new StringTokenizer(br.readLine());
 			int num = Integer.parseInt(st.nextToken());
 			for (int n = 0; n < num; n++) {
-				arr[i] += Integer.parseInt(st.nextToken());;
+				arr[i] += Long.parseLong(st.nextToken());;
 			}
 		}
 		Arrays.sort(arr);
-		int result = 0;
+		long result = 0;
 		for (int i = 0; i < N; i++) {
 			result += arr[i] * (N - i);
 		}
